Move BasisEvaluator to new class
