+++ tests/CMakeLists.txt
@@ -36,6 +36,7 @@ set (TEST_SOURCES
      tests/Test_Quadrature.cpp
      tests/Test_MonotoneComponent.cpp
      tests/Test_MultivariateExpansion.cpp
+     tests/Test_BasisEvaluator.cpp
      tests/Test_MultivariateExpansionWorker.cpp
      tests/Test_ArrayConversions.cpp
      tests/Test_ConditionalMapBase.cpp
